Splitting the ingredients and method sections into separate tabs
Ingredients and method in separate tabs. JS to make only method or only ingredients appear according to which button is clicked.

Implementation only on first recipe.

Issues to resolve: when implemented on all recipes, it will result in all ingredients or method tabs being opened when the button is clicked on only a particular recipe.